S-parameter calculation helper functions (#296)
* feat: initial sparams sim setup helper

* feat: extend sim through pml

* added extend_pml functinonality to sparam simulation builder

* feat: initial sparam calculation helper (non-functional)

* feat: index function in ObjectContainer

* feat: source recalculates on time steps on apply

* fix: minor fixes to sparams helper

* functional sparam calculation function

* fix: minor aset bug

* added exhaustive test cases

* feat: calculate full s-matrix

* cicd: more formatting rules

* cicd: added full ruff formatting rules

* fix: minor fixes

* cicd: Added RUF002 checks

* cicd: removed ruff exceptions

* fix: sparams test and string issue

* fix: source normalization position

* fix: center at

+++ src/fdtdx/colors.py
@@ -112,7 +112,7 @@ class Color(TreeClass):
         Returns:
             Tuple of (r, g, b) with integer values in [0, 255]
         """
-        return (int(round(self.r * 255)), int(round(self.g * 255)), int(round(self.b * 255)))
+        return (round(self.r * 255), round(self.g * 255), round(self.b * 255))
 
     def to_hex(self) -> str:
         """Return color as hexadecimal string.
